+++ .eslintignore
@@ -6,4 +6,4 @@ packages/react-scripts/fixtures
 fixtures/
 coverage/
 commitlint.config.js
-src/
+src
